States move to game logic
