+++ src/game/menus/gameover.rs
@@ -1,7 +1,7 @@
 use bevy::{prelude::*, input::mouse::MouseButtonInput};
 
 use crate::{
-    engine::states::GameState,
+    game::states::GameState,
     //globals::CHAR_SIZE,
     engine::asset_loaders::GraphicsAssets, 
     //render::ascii::spawn_ascii_text
