+++ src/game/player/mod.rs
@@ -9,7 +9,7 @@ pub use cursor::Cursor;
 
 use self::player_systems::{player_input, camera_follow, exit_step_check, player_mouse_input};
 
-use crate::engine::states::GameState;
+use crate::game::states::GameState;
