+++ src/main.rs
@@ -18,7 +18,7 @@ mod vectors;
 use game::GamePlugin;
 use engine::EnginePlugin;
 //use engine::states::{AppState, GameState, EngineState};
-use engine::states::{GameState, EngineState};
+use game::states::{GameState, EngineState};
 use globals::{HEIGHT, RESOLUTION, CLEAR};
